+++ src/Example/AnnotatedEntityExampleTypecastHandler.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Sirix\Cycle\Extension\Example;
 
 use Sirix\Cycle\Extension\Typecast\Chronos\ChronosToTimestampType;
+use Sirix\Cycle\Extension\Typecast\Handler\TypecastHandler;
 use Sirix\Cycle\Extension\Typecast\Uuid\UuidToStringType;
-use Vjik\CycleTypecast\TypecastHandler;
 
 class AnnotatedEntityExampleTypecastHandler extends TypecastHandler
 {