+++ test/Typecast/Array/ArrayToDelimitedStringTypeTest.php
@@ -7,8 +7,8 @@ namespace Sirix\Cycle\Extension\Test\Typecast\Array;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use Sirix\Cycle\Extension\Typecast\Array\ArrayToDelimitedStringType;
-use Vjik\CycleTypecast\CastContext;
-use Vjik\CycleTypecast\UncastContext;
+use Sirix\Cycle\Extension\Typecast\Context\CastContext;
+use Sirix\Cycle\Extension\Typecast\Context\UncastContext;
 
 final class ArrayToDelimitedStringTypeTest extends TestCase
 {