+++ test/Typecast/Uuid/UuidToBytesTypeTest.php
@@ -8,8 +8,8 @@ use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use Ramsey\Uuid\Uuid;
 use Sirix\Cycle\Extension\Typecast\Uuid\UuidToBytesType;
-use Vjik\CycleTypecast\CastContext;
-use Vjik\CycleTypecast\UncastContext;
+use Sirix\Cycle\Extension\Typecast\Context\CastContext;
+use Sirix\Cycle\Extension\Typecast\Context\UncastContext;
 
 final class UuidToBytesTypeTest extends TestCase
 {